feat(x11): Implement all remaining X11 requests
This change implements all remaining unimplemented X11 requests in `go/internal/x11/request_messages.go`. It also includes the following:

- Creation of `go/internal/x11/types.go` to define strongly-typed aliases for X11 resource identifiers and other protocol data structures.
- Refactoring of the entire `request_messages.go` file and its consumers to use these new strong types.
- Addition of reply message structs in `go/internal/x11/reply_messages.go` for requests that expect a response.
- Update of the `X11FrontendAPI` interface and its mock and WASM stub implementations to include methods for the newly supported requests.
- Addition of comprehensive tests to `go/internal/x11/request_messages_test.go` to cover the parsing of the newly implemented requests.
- Implementation of the `case` statements in `handleRequest` in `go/internal/x11/x11.go` to handle the newly parsed request types.

+++ go/internal/x11/x11_frontend_wasm.go
@@ -529,7 +529,7 @@ func (w *wasmX11Frontend) CreateWindow(xid xID, parent, x, y, width, height, dep
 		resizeMouseUp:   resizeMouseUp,
 	}
 	if values.Colormap != 0 {
-		w.windows[xid].colormap = xID{client: xid.client, local: values.Colormap}
+		w.windows[xid].colormap = xID{client: xid.client, local: uint32(values.Colormap)}
 	}
 
 	parentDiv.Call("appendChild", windowDiv)
@@ -2064,7 +2064,7 @@ func (w *wasmX11Frontend) ChangeWindowAttributes(xid xID, valueMask uint32, valu
 	if winInfo, ok := w.windows[xid]; ok {
 		style := winInfo.div.Get("style")
 		if valueMask&CWColormap != 0 {
-			winInfo.colormap = xID{client: xid.client, local: values.Colormap}
+			winInfo.colormap = xID{client: xid.client, local: uint32(values.Colormap)}
 		}
 		if valueMask&CWBackPixel != 0 {
 			r, g, b := w.GetRGBColor(winInfo.colormap, values.BackgroundPixel)
@@ -2077,7 +2077,7 @@ func (w *wasmX11Frontend) ChangeWindowAttributes(xid xID, valueMask uint32, valu
 			style.Set("borderColor", borderColor)
 		}
 		if valueMask&CWCursor != 0 {
-			w.SetWindowCursor(xid, xID{client: xid.client, local: values.Cursor})
+			w.SetWindowCursor(xid, xID{client: xid.client, local: uint32(values.Cursor)})
 		}
 	}
 	w.recordOperation(CanvasOperation{
